+++ src/lib.rs
@@ -1,6 +1,6 @@
 extern crate lazy_static;
 
-pub mod api;
+pub mod models;
 pub mod env;
 pub mod model;
 pub mod utils;
